requirements.txt
langchain
huggingface_hub
streamlit
python-dotenv